# Conflicts:
#	phpstan-baseline.neon

+++ src/PhpSpreadsheet/Reader/Slk.php
@@ -411,7 +411,7 @@ class Slk extends BaseReader
 
                     break;
                 case 'M':
-                    $formatArray['font']['size'] = substr($rowDatum, 1) / 20;
+                    $formatArray['font']['size'] = ((float) substr($rowDatum, 1)) / 20;
 
                     break;
                 case 'L':
